django/backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<5
django-filter>=24.3,<25

# Auth
dj-rest-auth[with_social]>=7.0,<8
djangorestframework-simplejwt>=5.3,<6
django-allauth>=65.0,<66

# Stripe
dj-stripe>=2.9,<3

# Celery
celery>=5.4,<6
django-celery-beat>=2.7,<3
redis>=5.0,<6

# Database
psycopg[binary]>=3.2,<4

# Logging
structlog>=24.4,<25
django-structlog>=8.1,<9

# Utils
django-extensions>=3.2,<4
python-decouple>=3.8,<4

# Email
django-anymail[sendgrid]>=12.0,<13

# Integrations
cryptography>=43.0,<44
requests>=2.32,<3
twilio>=9.0,<10
